Fix MSVC failure on HLSL files included in projects
The CMake MSVC generator will automatically attempt to compile any file with an `hlsl` extension.  The addition of the shaders to these projects are strictly for convenience and not meant to trigger any kind of build action against them.

Fixing this required distinguishing between GLSL and HLSL shaders, and adding a line to disable the additional tooling in MSVC environments